Read SHUTDOWN_TIMEOUT env var for the shutdown timeout value.
